fix: address PR review comments for jOOQ projection base
Fixes from Copilot code review:
- Fix #2: Change default jOOQ DB port from 62818 to 5432
- Fix #3: Fix connection leak by adding withTenantDsl() pattern
- Fix #4: Add column comments to jooq-init.sql for documentation
- Fix #5: Add !! to version field for null safety consistency
- Fix #6: Move slf4j-simple version to version catalog
- Fix #7: Make offset usage consistent (use Long directly)

All 13 integration tests pass.

+++ dvmm/dvmm-infrastructure/src/main/kotlin/de/acci/dvmm/infrastructure/projection/BaseProjectionRepository.kt
@@ -48,7 +48,7 @@ public abstract class BaseProjectionRepository<T : Any>(
         // Fetch the page using a wrapper query with LIMIT/OFFSET
         val items = dsl.selectFrom(query.asTable("subquery"))
             .limit(pageRequest.size)
-            .offset(pageRequest.offset.toInt())
+            .offset(pageRequest.offset)
             .fetch()
             .map { record -> mapRecord(record) }
